projet-flask-complet/requirements.txt
Flask>=2.0
google-generativeai
python-dotenv
requests  